requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
psycopg2-binary==2.9.9
requests==2.31.0
urllib3==2.2.1